Create GCD of 'n' integers - iterative code
To find the greatest common divisor of an n number of integers, take the integer at the 0th position, find GCD of this integer and the next integer. Take this GCD and find the GCD of this number and the next integer in the array. and so on till the last integer.